workflows_api/runtime/requirements.txt
# Waiting for https://github.com/stac-utils/stac-pydantic/pull/116 and 117
ddbcereal==2.1.1
fastapi>=0.75.1
fsspec==2023.3.0
mangum>=0.15.0
orjson>=3.6.8
psycopg[binary,pool]>=3.0.15
pydantic_ssm_settings>=0.2.0
pydantic>=1.9.0
pypgstac==0.7.4
python-multipart==0.0.5
requests>=2.27.1
s3fs==2023.3.0
stac-pydantic
xarray==2023.1.0
xstac==1.1.0
zarr==2.13.6
boto3==1.24.59
aws_xray_sdk>=2.6.0,<3
aws-lambda-powertools>=1.18.0
cryptography>=42.0.5
pyjwt>=2.8.0
